requirements.txt
azure-storage-blob==12.20.0
azure-search-documents==11.4.0
openai==1.30.1
tiktoken==0.7.0
streamlit==1.34.0